chore(server): move ExposedTestDb from util to utils
