+++ server/app/src/test/kotlin/pos/ambrosia/utest/BaseCurrencyTest.kt
@@ -4,7 +4,7 @@ import org.junit.After
 import org.junit.Before
 import pos.ambrosia.services.BaseCurrencyService
 import pos.ambrosia.services.CurrencyService
-import pos.ambrosia.util.ExposedTestDb
+import pos.ambrosia.utils.ExposedTestDb
 import java.io.File
 import kotlin.test.Test
 import kotlin.test.assertEquals
