+++ server/app/src/test/kotlin/pos/ambrosia/utest/ConfigServiceTest.kt
@@ -5,7 +5,7 @@ import org.junit.After
 import org.junit.Before
 import pos.ambrosia.models.Config
 import pos.ambrosia.services.ConfigService
-import pos.ambrosia.util.ExposedTestDb
+import pos.ambrosia.utils.ExposedTestDb
 import java.io.File
 import kotlin.test.Test
 import kotlin.test.assertEquals
