+++ server/app/src/test/kotlin/pos/ambrosia/utest/PermissionsServiceTest.kt
@@ -3,7 +3,7 @@ package pos.ambrosia.utest
 import org.junit.After
 import org.junit.Before
 import pos.ambrosia.services.PermissionsService
-import pos.ambrosia.util.ExposedTestDb
+import pos.ambrosia.utils.ExposedTestDb
 import java.io.File
 import java.util.UUID
 import kotlin.test.Test
