+++ server/app/src/test/kotlin/pos/ambrosia/utest/SpaceServiceTest.kt
@@ -5,7 +5,7 @@ import org.junit.After
 import org.junit.Before
 import pos.ambrosia.models.Space
 import pos.ambrosia.services.SpaceService
-import pos.ambrosia.util.ExposedTestDb
+import pos.ambrosia.utils.ExposedTestDb
 import java.io.File
 import kotlin.test.Test
 import kotlin.test.assertEquals
